Push or Not with token setting
