+++ scripts/script.py
@@ -2,7 +2,36 @@ import aiohttp
 import asyncio
 import json
 from datasets import Dataset
+from utils import transform_to_jsonl
 import jsonlines
+from argparse import ArgumentParser
+
+
+def argument_parser():
+    parser = ArgumentParser()
+    parser.add_argument(
+        "--input_file",
+        type=str,
+        default="word.txt",
+        help="Input file containing the list of words to fetch",
+    )
+    parser.add_argument(
+        "--output_file",
+        type=str,
+        default="../data/dataset.jsonl",
+        help="Output file to store the fetched data",
+    )
+    parser.add_argument(
+        "--delay", type=int, default=1, help="Delay between each request in seconds"
+    )
+    parser.add_argument(
+        "--push_to_hub", type=bool, default=True, help="Push the dataset to the hub"
+    )
+    parser.add_argument("--token", type=str, default="", help="Hugging Face API token")
+
+    parser.add_argument("--repo_id", type=str, default="wolof-french-dictionary", help="The repo id name")
+
+    return parser.parse_args()
 
 
 async def fetch_word_data(session, word):
@@ -17,57 +46,6 @@ async def fetch_word_data(session, word):
                 f"Erreur lors de la récupération des données pour le mot: {word} (Code: {response.status})"
             )
             return None
-
-
-def transform_to_jsonl(data):
-    transformed_data = []
-
-    if isinstance(data, dict):
-        lxGroups = data.get("lxGroup", [])
-
-        # Convert lxGroups to a list if it's a single object
-        if isinstance(lxGroups, dict):
-            lxGroups = [lxGroups]
-
-        for lxGroup in lxGroups:
-            if isinstance(lxGroup, dict):
-                psGroup = lxGroup.get("psGroup", {})
-                if isinstance(psGroup, dict):
-                    geGroup = psGroup.get("geGroup", {})
-                    if isinstance(geGroup, dict):
-                        xvGroup = geGroup.get("xvGroup", [])
-
-                        if isinstance(xvGroup, dict):
-                            xvGroup = [xvGroup]
-
-                        if isinstance(xvGroup, list):
-                            if lxGroup.get("lx", "") and geGroup.get("ge", ""):
-                                transformed_data.append(
-                                    {
-                                        "input": lxGroup.get("lx", ""),
-                                        "output": {
-                                            "definition": geGroup.get("ge", "")
-                                            if "ge" in geGroup
-                                            else "",
-                                            "etymology": lxGroup.get("etGroup", {}).get(
-                                                "et", ""
-                                            )
-                                            if "et" in lxGroup.get("etGroup", {})
-                                            else "",
-                                            "synonym": psGroup.get("sy", {})
-                                            if "sy" in psGroup
-                                            else "",
-                                        },
-                                    }
-                                )
-
-                            for entry in xvGroup:
-                                xv = entry.get("xv", "")
-                                xe = entry.get("xe", "")
-                                if xv and xe:
-                                    transformed_data.append({"input": xv, "output": xe})
-    print(f"Données transformées: {transformed_data}")  # Débogage
-    return transformed_data
 
 
 async def fetch_and_write_word_data(session, word, file, delay):
@@ -102,6 +80,7 @@ async def create_jsonl_from_words(input_file, output_file, delay=1):
 
     print(f"Les données ont été ajoutées avec succès dans le fichier: {output_file}")
 
+
 def process_jsonl_to_dataset(jsonl_file):
     french = []
     wolof = []
@@ -110,33 +89,30 @@ def process_jsonl_to_dataset(jsonl_file):
         data = list(reader)
 
     for line in data:
-        wolof.append(str(line['input']))
-        if line['output'] is not None:
-            if isinstance(line['output'], dict) and 'definition' in line['output']:
-                french.append(str(line['output']['definition']))
+        wolof.append(str(line["input"]))
+        if line["output"] is not None:
+            if isinstance(line["output"], dict) and "definition" in line["output"]:
+                french.append(str(line["output"]["definition"]))
             else:
-                french.append(str(line['output']))
+                french.append(str(line["output"]))
         else:
-            french.append(None)  
+            french.append(None)
 
     if len(french) != len(wolof):
-        raise ValueError('The number of french and wolof sentences are not equal')
+        raise ValueError("The number of french and wolof sentences are not equal")
 
     ds = Dataset.from_dict({"wolof": wolof, "french": french})
     return ds
 
+
 async def main():
-    input_file = 'word.txt'
-    output_file = '../data/dataset.jsonl'
-    delay = 1
 
-    await create_jsonl_from_words(input_file, output_file, delay)
+    args = argument_parser()
+    await create_jsonl_from_words(args.input_file, args.output_file, args.delay)
+    dataset = process_jsonl_to_dataset(args.output_file)
+    if args.push_to_hub:
+        dataset.push_to_hub(args.repo_id, token=args.token)
 
-    ds = process_jsonl_to_dataset(output_file)
-
-    dataset_dict = ds.train_test_split(test_size=0.2)
-
-    dataset_dict.push_to_hub("jolof")   # Push to the hub
 
 if __name__ == "__main__":
     asyncio.run(main())
